feat: add bulk endpoint for ingest api

+++ src/domains/ingest/model.rs
@@ -5,6 +5,6 @@ use uuid::Uuid;
 #[derive(Clone, Debug, FromRow)]
 pub struct Measurement {
     pub node_id: Uuid,
-    pub count: i32,
+    pub count: i64,
     pub created_at: Option<DateTime<Utc>>
 }
